+++ Inspector/Functions.m
@@ -157,7 +157,7 @@ NSString *fsDescription(unsigned long long size)
 	if (size == 0) {
 		sizeStr = @"0 bytes";
 	} else if (size < (10 * ONE_KB)) {
-		sizeStr = [NSString stringWithFormat:@"%s %d bytes", sign, size];
+		sizeStr = [NSString stringWithFormat:@"%s %d bytes", sign, (long)size];
 	} else if (size < (100 * ONE_KB)) {
  		sizeStr = [NSString stringWithFormat:@"%s %3.2fKB", sign,
                           					((double)size / (double)(ONE_KB))];
